requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==11.0.0
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.1.3
torch==2.1.1
torchvision==0.16.1
scipy==1.11.4
aiohttp==3.9.1
python-dotenv==1.0.0
pytest==7.4.3
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
yfinance==0.2.54
requests==2.32.3
joblib==1.3.2
h5py==3.11.0
tensorflow-cpu==2.16.1; python_version < "3.13"
